requirements.txt
pulumi==3.194.0
pulumi-azure-native==3.8.0
pulumi-azure==6.26.0
azure-cli==2.77.0
azure.identity==1.24.0
dpath==2.2.0
pyyaml==6.0.2
azure.mgmt.network==29.0.0